requirements.txt
fastapi
uvicorn
python-dotenv
watchdog
pydantic
httpx
aiohttp
python-multipart 